Renamed inputs to binds
